Renaming files, adding to index.
