+++ react/src/components/atoms/forms/InputCurrency/InputCurrency.knobs.options.js
