+++ react/src/components/atoms/forms/InputCurrency/InputCurrency.stories.js
@@ -2,15 +2,15 @@ import React from 'react';
 
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
-import { number, withKnobs } from '@storybook/addon-knobs/react';
+import { withKnobs } from '@storybook/addon-knobs/react';
 import { action } from '@storybook/addon-actions';
 
 import InputCurrency from './index';
-import InputTextOptions from './InputNumber.knobs.options';
+import InputCurrencyOptions from './InputCurrency.knobs.options';
 
 storiesOf('atoms/forms', module).addDecorator(withKnobs)
   .add('InputCurrency', withInfo('<div></div>')(() => {
-    const inputTextOptionsWithKnobs = Object.assign(...Object.entries(InputTextOptions).map(([k, v]) => (
+    const inputTextOptionsWithKnobs = Object.assign(...Object.entries(InputCurrencyOptions).map(([k, v]) => (
       { [k]: v() })));
     inputTextOptionsWithKnobs.onChange = action('Text input modified');
