+++ react/src/components/atoms/forms/InputCurrency/index.js
@@ -3,44 +3,8 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import numbro from 'numbro';
 
-import './style.css';
-import { InputContext } from './context';
-
-export class Input extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      value: this.props.defaultText,
-      updateState: (newState) => this.setState(newState),
-      showError: false,
-      errorMsg: this.props.errorMsg
-    };
-  }
-  render() {
-    const inputLabelClasses = classNames({
-      ma__label: true,
-      'ma__label--hidden': (this.props.labelText && this.props.hiddenLabel),
-      'ma__label--required': (this.props.labelText && this.props.required),
-      'ma__label--optional': (this.props.labelText && !this.props.required)
-    });
-    return(
-      <InputContext.Provider value={this.state}>
-        <React.Fragment>
-          {this.props.labelText && <label htmlFor={this.props.id} className={inputLabelClasses}>{this.props.labelText}</label>}
-          {this.props.children}
-          {this.state.showError && this.props.errorMsg.length > 0 && <div className="ma__error-msg">{this.state.errorMsg}</div>}
-        </React.Fragment>
-      </InputContext.Provider>
-    );
-  }
-}
-
-Input.defaultProps = {
-  hiddenLabel: false,
-  required: false,
-  readonly: false
-};
-
+import Input from '../Input';
+import { InputContext } from '../Input/context';
 
 const Currency = (props) => (
   <React.Fragment>
