+++ react/src/index.js
@@ -23,6 +23,10 @@ export InputText from './components/atoms/forms/InputText';
 export SelectBox from './components/atoms/forms/SelectBox';
 export InputTextTypeAhead from './components/atoms/forms/InputTextTypeAhead';
 export InputTextFuzzy from './components/atoms/forms/InputTextFuzzy';
+export Input from './components/atoms/forms/Input';
+// Also export the context.
+export { InputContext } from './components/atoms/forms/Input/context';
+export InputCurrency from './components/atoms/forms/InputCurrency';
 // @atoms/@headings
 export ColoredHeading from './components/atoms/headings/ColoredHeading';
 export CompHeading from './components/atoms/headings/CompHeading';
